+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.??.?? (2021-??-??)
+
+#### :boom: Breaking Change
+
+* Renamed `noMessageBodyStatusCodes` to `noContentStatusCodes` `core/request/response`
+
 ## v3.43.0 (2021-04-14)
 
 #### :house: Internal